Rename tests script
The script to run tests can be used locally or on docker.

By default, the script will run with your local python setup. You
can tell it otherwise running the script with the `virtualenv`
parameter:

```
  $> sh ./run-tests.sh virtualenv
```

It will create a virtualenv directory for you located in
`${VIRTUALENV_DIR}` which defaults to `venv`.