Work around for Python 3.9's GC behavior wrt memory views
Python 3.9's behavior WRT garbage collecting memoryviews causes
segmentation faults where cycles aren't freed in the correct order,
causing corruption and even double-frees when sharedbuffers is
holding onto a complex graph of related memoryviews.

Work around this behavior by actively breaking cycles early
in the ProxyBufferObject's finalizer. Simply clearing the idmap
is enough to break the cycles early, letting the normal refcount
mechanism do most of the work and avoiding the bugs.

This looks like a bug in python and not sharedbuffers TBH,
but this workaround seems to be effective enough.